업데이트
효과음 추가, 스킬 리소스 추가, 맵 수정, 보스 수정, 원거리 투사체(적) 추가 및 공격 사거리 수정

+++ Assets/Game/Runtime/Core/PrototypeGameSession.RunFlow.cs
@@ -53,7 +53,7 @@ namespace SimpleGame
 
         public void ShowDamagePopup(
             Vector3 worldPosition,
-            float amount,
+            int amount,
             DamagePopupStyle style)
         {
             combatFeedback?.ShowDamagePopup(
